# backend/services/ridely-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ridely Service",
        "description": "Provide Taxi services",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/drivers/{id}": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Buscar um motorista",
                "operationId": "1d3ed74a84af9e876a1b2e9662e61574",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do motorista",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Motorista encontrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriverShowSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Corrida não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseRideNotFound"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no serviço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInternalError"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Serviço indisponivel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseServiceUnavailable"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Remove um motorista",
                "operationId": "3670a9b76c6831f3c02d896166bb485a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do motorista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Motorista removido com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidParams"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseUnauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no serviço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInternalError"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Serviço indisponivel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseServiceUnavailable"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/drivers": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Lista os motoristas com filtros opcionais",
                "operationId": "7f3fc5cabb720d907c416cbf56c91826",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Referente a qual página dos itens da listagem deseja buscar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade de resultados por página (máx: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Campo para ordenação (ex: name, created_at)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Direção da ordenação: asc ou desc",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Campos específicos a retornar (ex: name,available)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de motoristas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriverListSuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidParams"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseUnauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no serviço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInternalError"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Serviço indisponivel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseServiceUnavailable"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Cria um novo motorista",
                "operationId": "616886d5a7effb20d218264e6804538a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "car"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "car": {
                                        "$ref": "#/components/schemas/Car"
                                    },
                                    "available": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Motoristas criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriverCreatedSuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidParams"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseUnauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no serviço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInternalError"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Serviço indisponivel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseServiceUnavailable"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/drivers/{id}/get-rides": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Lista corridas abertas para um motorista",
                "operationId": "1e91ae6a7e7f76ddc842d69403e6cd09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do driver",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de rides abertas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriverGetRidesSuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidParams"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhuma corrida aberta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseNoRidesWaiting"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseUnauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no serviço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInternalError"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Serviço indisponivel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseServiceUnavailable"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Car": {
                "required": [
                    "license_plate",
                    "model",
                    "color"
                ],
                "properties": {
                    "license_plate": {
                        "type": "string",
                        "example": "ABC1234"
                    },
                    "model": {
                        "type": "string",
                        "example": "Fiat Uno"
                    },
                    "color": {
                        "type": "string",
                        "example": "Vermelho"
                    }
                },
                "type": "object"
            },
...
            "DriverShowSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "label": {
                        "type": "string",
                        "example": "success"
                    },
                    "code": {
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Driver"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponseInternalError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "label": {
                        "type": "string",
                        "example": "common.error.internal_error"
                    },
                    "code": {
                        "type": "integer",
                        "example": 20
                    },
                    "message": {
                        "type": "string",
                        "example": "Internal server error"
                    },
                    "params": {
                        "description": "Parametros da requisição",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "id"
                        }
                    },
                    "detail": {
                        "description": "Stack trace (visível apenas em ambiente de desenvolvimento)",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ErrorResponseRideNotFound": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "label": {
                        "type": "string",
                        "example": "common.error.ride_not_found"
                    },
                    "code": {
                        "type": "integer",
                        "example": 61
                    },
                    "message": {
                        "type": "string",
                        "example": "Ride not found"
                    },
                    "params": {
                        "description": "Parametros da requisição",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "id"
                        }
                    },
                    "detail": {
                        "description": "Stack trace (visível apenas em ambiente de desenvolvimento)",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ErrorResponseUnauthorized": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "label": {
                        "type": "string",
                        "example": "common.error.invalid_token"
                    },
                    "code": {
                        "type": "integer",
                        "example": 25
                    },
                    "message": {
                        "type": "string",
                        "example": "Invalid token: token is expired"
                    },
                    "params": {
                        "description": "Parametros da requisição",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "id"
                        }
                    },
                    "detail": {
                        "description": "Stack trace (visível apenas em ambiente de desenvolvimento)",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}